Split plan value shapes
- Move function value shapes into a dedicated plan value child module.
- Move capture value shapes into a dedicated plan value child module.
- Keep value.rs as the ValueType, FunctionType, Value, and ListValue facade.
- Preserve existing crate::plan re-exports and colocate unit tests with their owning modules.
